+++ experimental/CHANGELOG.md
@@ -10,6 +10,8 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 
 ### :rocket: Features
 
+* feat(configuration): add resource detection parsing [#XXXX](https://github.com/open-telemetry/opentelemetry-js/pull/XXXX) @MikeGoldsmith
+
 ### :bug: Bug Fixes
 
 * fix(otlp-exporter-base): remove sendBeacon in favor of fetch with keepalive [#6391](https://github.com/open-telemetry/opentelemetry-js/pull/6391) @overbalance